requirements.txt
huggingface_hub
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.5
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
datasets==4.0.0
dill==0.3.8
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.3.0
gdown==5.2.0
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
intel-openmp==2021.4.0
Jinja2==3.1.6
kiwisolver==1.4.9
lxml==6.0.1
MarkupSafe==3.0.2
matplotlib==3.10.6
mkl==2021.4.0
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
networkx==3.4.2
numpy==1.26.4
onnx==1.16.1
onnxruntime-gpu==1.22.0
onnxslim==0.1.34
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.2
pillow==11.3.0
propcache==0.3.2
protobuf==6.32.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyparsing==3.2.3
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
soupsieve==2.8
sympy==1.14.0
tbb==2021.11.0
torch==2.3.1+cu121
torchaudio==2.3.1+cu121
torchvision==0.18.1+cu121
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.0
ultralytics-thop==2.0.17
urllib3==2.5.0
xlsxwriter==3.2.5
xxhash==3.5.0
yarl==1.20.1
